Haiku: move BSD API to a separate file
Haiku implements various parts of the (non-POSIX) BSD API. This moves it to a
separate file, for easier future maintenance.

No functional change intended; the changes are synchronized with R1Beta5
